refactor : component 접미사 제거
